isolate BMI class
