+++ app/src/main/java/com/corposense/testingandroid/MainActivity.java
@@ -41,22 +41,7 @@ public class MainActivity extends AppCompatActivity {
                 tvResult.setText( "Your BMI = " + bmi.calculate().toString()+ " kg/m2" );
             }
         });
-
     }
 
 
-    public static class BodyMassIndex {
-
-        private double weight;
-        private double height;
-
-        public BodyMassIndex(double weight, double height) {
-            this.weight = weight;
-            this.height = height;
-        }
-
-        public Double calculate(){
-            return this.weight / (this.height * this.height);
-        }
-    }
 }
